Fix build (#86)
* add key to deploy for "test image"

* Move deploy action to SDR-E common action

* update on pr action to use sdre-common

Move to common github action.

This removes the dependency on "change files" and instead goes to explicitly state that only files in `rootfs` and the `dockerfile` will trigger this action.

* Fix typo in readme

Linter caught it, otherwise I would've never found it :)

* Fix markdown linting

Disabling the check on unordered list indentation....for now.

* don't exit 1 on ais upstream check

Instead, we'll set a variable for the next job to check and only run if
there is an update. Exit status other than 0 is now reserved for actual
fail states